+++ exporters/stats/stackdriver/src/test/java/io/opencensus/exporter/stats/stackdriver/StackdriverExportUtilsTest.java
@@ -227,9 +227,9 @@ public class StackdriverExportUtilsTest {
         .isEqualTo(
             com.google.protobuf.Timestamp.newBuilder().setSeconds(100).setNanos(3000).build());
 
-    // Proto timestamp doesn't allow negative values, instead it will replace the negative values
+    // Stackdriver doesn't allow negative values, instead it will replace the negative values
     // by returning a default instance.
-    Timestamp censusTimestamp2 = Timestamp.create(-100, -3000);
+    Timestamp censusTimestamp2 = Timestamp.create(-100, 3000);
     assertThat(StackdriverExportUtils.convertTimestamp(censusTimestamp2))
         .isEqualTo(com.google.protobuf.Timestamp.newBuilder().build());
   }